Mer strukturert feilresponser (#624)

+++ app/main/api/Maksimum.kt
@@ -72,7 +72,7 @@ data class UtbetalingMedMer(
 
 data class Reduksjon(
     val timerArbeidet: Double,
-    val annenReduksjon: Float
+    val annenReduksjon: Double
 )
 
 data class Periode(
